Release Candidate 1 : highly optimized and fixed most issues, https://github.com/n-ce/ytify/issues/59#issuecomment-1338861692 for more details
