--- CSS/player.css
@@ -1,75 +1,0 @@
-
-#playerContainer {
-  height: auto;
-  width: auto;
-  padding: 0.3rem;
-
-  background: var(--accent);
-  border: 0.1px solid var(--border);
-  box-shadow: 0 0.1rem 0.3rem #0003;
-  border-radius: 0.7rem;
-
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: space-between;
-  animation: fade 500ms ease-out;
-}
-
-#playbackButtons {
-  width: auto;
-  display: flex;
-}
-
-#playButton {
-  height: 3rem;
-  width: 3rem;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  border-radius: 50%;
-}
-#playSpeed{
-  color: var(--border);
-  border: none;
-  background: none;
-  outline:none;
-}
-#settingsButton {
-  transition: all 1s;
-}
-
-#queuePreviousButton {
-  transform: scaleX(-1)
-}
-
-input {
-  height: 3rem;
-  width: 18rem;
-  margin-top: 1rem;
-
-  background: var(--accent);
-  border: 1px solid var(--border);
-  border-radius: 0.5rem;
-  box-shadow: 0 0.1rem 0.3rem #0003;
-  color: var(--text);
-
-  outline: none;
-  text-align: center;
-  font-weight: bold;
-  animation: fade 1s ease-out
-}
-
-@media (min-width:600px) {
-  #playerContainer {
-    height: 3rem;
-    width: auto;
-    flex-direction: row;
-    justify-content: space-between;
-  }
-
-  input {
-    margin-top: 0;
-  }
-}
-
