+++ Stylesheets/player.css
@@ -1,0 +1,96 @@
+#playerContainer {
+  height: auto;
+  width: auto;
+  padding: 0.3rem;
+  background: var(--accent);
+  border: 0.1px solid var(--border);
+  box-shadow: 0 0.1rem 0.3rem #0003;
+  border-radius: 0.7rem;
+  animation: fade 500ms ease-out;
+}
+
+#playerControls {
+  display: flex;
+  justify-content: space-between
+}
+
+#playSpeed {
+  color: var(--border);
+  border: none;
+  background: none;
+  outline: none;
+}
+
+#playerProfiles {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+}
+
+[data-badge]:after {
+  font-family: 'Noto Sans';
+  content: attr(data-badge);
+  font-size: 1rem;
+  transition: background 0.5s, color 1s;
+}
+
+[data-badge^="-"]:after,
+[data-badge=""]:after,
+[data-badge="0"]:after {
+  display: none;
+}
+
+
+#settingsButton {
+  transition: all 0.5s ease-in;
+}
+
+input[type="url"] {
+  height: 3rem;
+  width: 9rem;
+
+  background: var(--accent);
+  border: 1px solid var(--border);
+  border-radius: 0.5rem;
+  box-shadow: 0 0.1rem 0.3rem #0003;
+  color: var(--text);
+
+  font-family: 'Noto Sans';
+  outline: none;
+  text-align: center;
+  font-weight: bold;
+  animation: fade 1s ease-out
+}
+
+@media (min-width:600px) {
+  #playerContainer {
+    display: flex;
+  }
+}
+
+
+input[type="range"] {
+  -webkit-appearance: none;
+  background-color: transparent;
+  width: 10rem;
+  cursor: pointer;
+}
+
+input[type="range"]::-webkit-slider-runnable-track {
+  -webkit-appearance: none;
+  width: 100%;
+  height: 0.5rem;
+  background-color: var(--accent);
+  border-radius: 1rem;
+  border: 0.1rem solid var(--border);
+}
+
+input[type="range"]::-webkit-slider-thumb {
+  -webkit-appearance: none;
+  height: 1rem;
+  width: 1rem;
+  border-radius: 1rem;
+  background-color: var(--accent);
+  border: 0.1rem solid var(--border);
+  margin-top: -0.35rem
+}
